+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 2.0.0a1 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Added Plone 6.1 version in buildout.
+  [chris-adam]
 
 
 2.0.0a (2023-11-28)